Adding separate test scripts for error injection verfification using log and dcgm callbacks
